perf(qwen): run the stream in bfloat16 through load, noise and every streamed pass
Qwen-Image ran its whole transformer in float32 by accident, through three
leaks, any one of which was enough: the noise came out of MLXRandom.normal
in float32; the packer's float32 scales and biases were loaded as they were,
and MLX's quantized matmul takes its output dtype from them; and on a streamed
load LayerWeightStream.release handed every slot the raw node from loadArrays,
so the second step on read float32 scales even where the first had not. At a
4096-token image a float32 stream also puts the attention off the fused
kernel, which is where the seconds went.

- QwenImageTransformerPrecision is the family's activation dtype, bfloat16 by
  default and float32 under ZEPHRA_DIT_DTYPE=f32, seeded from the environment
  for now exactly as klein's is; the "statics from env" item turns both into
  a value the backend passes in.
- QwenImagePipeline.generate casts the packed noise and the conditioning to
  it before the loop. The reference mix already follows the noise's dtype and
  the Euler step keeps the sample's, so the SDEdit path inherits it.
- QwenImageTransformer casts its text to the latents' dtype at entry, beside
  the conditioning cast already there. The transformer follows its inputs the
  way klein's does rather than casting to the static itself, so the float32
  parity fixtures run as they are, in float32, without an environment
  variable. That is the one place this departs from the design's wording
  ("casts inputs at entry"): the effect is the same in production, and the
  tests then set what they need.
- QwenImageWeightLoading.castFloatParameters casts the float32 parameters of
  the text encoder and the transformer at load, before either
  LayerWeightStream is attached, so the streams capture the cast nodes. A
  copy of klein's for now; M8 moves both into one shared loader. The
  autoencoder stays float32 on purpose.
- LayerWeightStream.Slot records the dtype the tree held at capture and
  release casts each fresh node back to it: a lazy cast over the lazy read,
  evaluated by the next prefetch, kilobytes of scales per layer.

Tests: LayerWeightStreamTests "a slot cast at load keeps its dtype on every
later pass" (two passes, every parameter bfloat16 after each, output equal to
a resident bfloat16 pass); TransformerParityTests "a bfloat16 stream stays
bfloat16 through a block" (a packed doll's-house block with float32 scales
widens the stream, the cast one does not) and "the whole stack runs in
whatever dtype its latents arrive in"; SchedulerTests "a step keeps the
sample's dtype". The existing float32 parity fixtures still pass unchanged.

Docs: AGENTS.md "Streaming the weights" and "Third model", README's
Qwen-Image section. The catalog's measured figures are left as they are and
marked due a rerun on an idle machine.

+++ Packages/QwenImageKit/Sources/QwenImage/Model/Transformer/QwenImageTransformer.swift
@@ -54,6 +54,10 @@ public final class QwenImageTransformer: Module {
     ///   - timestep: The noise level, from zero to one.
     ///   - frequencies: Rotary tables for both streams, sized to this image and prompt.
     ///
+    /// The stream runs in the latents' dtype, which the pipeline chooses through
+    /// `QwenImageTransformerPrecision`; the text and the conditioning are cast to it here so
+    /// neither can widen it. The sinusoid itself is built in float32 whatever arrives.
+    ///
     /// Throws only when streaming: a shard that changed under the model, or a stop between
     /// blocks. A resident run has nothing to throw.
     public func callAsFunction(
@@ -63,7 +67,7 @@ public final class QwenImageTransformer: Module {
         frequencies: (image: RotaryFrequencies, text: RotaryFrequencies)
     ) throws -> MLXArray {
         var image = imageInput(latents)
-        var textStream = textInput(textNorm(text))
+        var textStream = textInput(textNorm(text.asType(latents.dtype)))
         let conditioning = timeEmbedding(timestep).asType(image.dtype)
 
         let step = { (block: QwenImageTransformerBlock) in
